requirements.txt
requests
python-dateutil
cfscrape